#1265: replay: updated filenames to not ref stats
